kbfs: enable stricter golangci linting
And pin the linter to a particular version.

Suggested by joshblum.

Issue: #18565

+++ go/kbfs/libkbfs/disk_limits_unix.go
@@ -28,14 +28,14 @@ func getDiskLimits(path string) (
 	}
 
 	// Bavail is the free block count for an unprivileged user.
-	availableBytes = uint64(stat.Bavail) * uint64(stat.Bsize)
-	totalBytes = uint64(stat.Blocks) * uint64(stat.Bsize)
+	availableBytes = stat.Bavail * uint64(stat.Bsize)
+	totalBytes = stat.Blocks * uint64(stat.Bsize)
 	// Some filesystems, like btrfs, don't keep track of inodes.
 	// (See https://github.com/keybase/client/issues/6206 .) Use
 	// the total inode count to detect that case.
 	if stat.Files > 0 {
-		availableFiles = uint64(stat.Ffree)
-		totalFiles = uint64(stat.Files)
+		availableFiles = stat.Ffree
+		totalFiles = stat.Files
 	} else {
 		availableFiles = math.MaxInt64
 		totalFiles = math.MaxInt64
